pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "simple_fastapi"
version = "0.1.0"
description = "A simple FastAPI project"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi[standard]",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
]

[tool.hatch.build.targets.wheel]
packages = ["src/simple_fastapi"]

[tool.hatch.build.targets.sdist]
include = ["src"]

